tools/requirements.txt
flask
flask-cors
pydub
SpeechRecognition
ffmpeg-python
python-dotenv
Werkzeug
google-generativeai
